Generate a cloud of particles from a multivariate normal distribution

Takes the mean and variance for a multivariate normal distribution, as well as the number of particles to generate and return random draws from the multivariate normal if the numbers of particles is > 0, otherwise return NULL. At least one of mean or sigma must be provided.

particle_draws(n, mu, covar)

Arguments

n
number of observations

mu
mean vector

covar
covariance matrix

Value

If n > 0 returns n draws from the multivariate normal with mean and sigma, otherwise returns NULL

Examples

pmwg:::particle_draws(100, rep(0.2, 7), diag(rep(7)))

#>               [,1]         [,2]         [,3]        [,4]         [,5]
#>   [1,] -0.52754568 -0.000998021 -0.321967022 -1.00436543  1.257657630
#>   [2,]  0.25682579 -0.120208324 -0.866683691  1.09783313  2.170578992
#>   [3,]  0.98513372 -0.131784477 -2.313380995 -0.31020926  0.708839517
#>   [4,]  0.26797073  1.051084479 -0.495165276  0.12233936  1.023274183
#>   [5,] -0.15699135 -0.216969694  3.196007067  3.74044959 -0.039865886
#>   [6,] -1.28524979  1.324077819  0.127923921 -0.59409025  1.492878584
#>   [7,]  0.01665405  1.473770599 -0.200109239  1.37238850 -1.291496088
#>   [8,]  0.76783817  0.652275479 -1.844582633 -1.06749687  0.397388616
#>   [9,] -1.19224950  1.168670020  0.679376076 -0.75695479 -0.373222832
#>  [10,]  0.12416857  0.976105283  0.722542104 -0.37688730 -0.921226524
#>  [11,] -0.49706045 -0.572487726 -0.060035359 -2.48739289  0.808492721
#>  [12,]  0.12362202  0.977345878  0.247076244  0.20208305 -0.132607999
#>  [13,] -1.54015331  1.884944066 -0.868124137  1.36151026  0.894337760
#>  [14,]  0.61339410  1.405202040  0.475647476 -1.46321478  0.131346489
#>  [15,]  0.23826375  0.128207641  0.995577706  1.37499138 -0.008685087
#>  [16,]  0.23528835  0.300999190 -1.044668775  0.96830034  0.977793161
#>  [17,]  0.14511858  0.344770279  1.848431308 -0.78500152 -0.669747825
#>  [18,] -0.59860348  1.847979122  0.290489175  0.35951436  0.167531961
#>  [19,]  0.44309145 -0.689184542 -0.279791502  0.09574278  2.054624913
#>  [20,]  1.76601079 -0.233535061 -1.137820789  1.63095955 -0.434594611
#>  [21,]  0.18932001  0.657366881 -0.148856205  0.43028960  2.128505058
#>  [22,]  1.57427481 -0.558206986 -0.973454775  1.39676287  0.295587758
#>  [23,]  0.10650682 -1.966843759 -0.838786007 -1.03287994 -0.605737851
#>  [24,] -0.29308072  0.476567973  0.650052826  0.19707040  0.259082911
#>  [25,] -0.56325233 -1.225261728  0.097271068 -0.28661881 -0.341655923
#>  [26,] -1.42098593  1.481376044  1.287230840 -0.01808515  0.632795587
#>  [27,]  0.95564050 -0.221960367  0.485425141  0.49234421  0.558313127
#>  [28,]  0.80003544 -0.898974861 -0.894889838 -0.41144685 -0.299767797
#>  [29,] -0.61283343 -0.250467378 -1.209313171 -0.03815355 -0.046157526
#>  [30,]  0.26441535 -0.382849233  0.713731055 -1.20225565  0.564407506
#>  [31,]  0.19537769 -1.854510823  0.714849232  0.70221836  2.716059079
#>  [32,] -0.67406533  0.622000744 -0.084728679  0.83847327  0.107295727
#>  [33,] -0.60762293 -1.080136789  1.107162166  0.92926918 -0.007270708
#>  [34,] -0.25613369  1.191935861  0.898148579 -0.36505799  0.978285503
#>  [35,]  0.91099258  0.154646131  0.260164017 -2.21173676  1.014730628
#>  [36,] -0.66173744  1.359878850  0.601906735 -0.55634134 -0.025517760
#>  [37,] -1.28235650  0.234785243 -1.919542090  1.32340962 -0.460076775
#>  [38,] -0.94778165  0.391373541  0.485754571  0.84932018 -2.413990787
#>  [39,] -0.64870541 -1.129538482  0.256908179  1.25157942 -1.030019437
#>  [40,]  2.73336643  0.028053259  0.647195998  0.57502720  0.570709728
#>  [41,]  0.10167357 -0.336840373  0.980407133 -0.04635319  2.995608963
#>  [42,] -0.24488789  0.319561819  2.034395451  1.87999494 -1.334051594
#>  [43,]  0.80292388  0.351939306  0.005459749  1.33273707 -0.076147993
#>  [44,] -0.29426926  0.771986137  1.131802994 -0.79840813  0.388068676
#>  [45,]  1.21445358  0.296078674 -0.415914250  2.03438743 -0.139851837
#>  [46,] -1.78740603  1.589361220  0.161641134  0.74589853 -0.861263262
#>  [47,]  1.20776433  0.584007467 -1.522262731  3.27795256  2.025567156
#>  [48,] -2.02941307  1.561257119  2.226705048  0.81189553 -0.591587905
#>  [49,] -1.68289757 -1.370608238  1.364605262 -0.42869759  0.727338031
#>  [50,] -0.46892084 -0.234609848  0.216868684 -0.41043000  0.178757313
#>  [51,]  0.49479438 -0.253393905 -0.638124907  0.56329855  1.665152034
#>  [52,] -0.93231041  1.515448566 -0.618602346  0.01530779  0.434292652
#>  [53,] -0.32460740  0.011778890  0.451585655 -0.64399762  2.425045127
#>  [54,] -0.25902132 -0.884122045 -0.298652546  0.61730698 -0.765463617
#>  [55,] -1.38970840  0.620069865 -0.410949464  0.68323567 -0.304945488
#>  [56,] -0.10684518  1.757381967 -1.088019804  0.89689319 -0.446489875
#>  [57,]  0.80640447  2.602792384  0.157169391  2.11312409 -0.262156877
#>  [58,]  0.63169954  1.390711937  0.013484808 -1.33581363 -1.072669274
#>  [59,]  1.33256297  0.843061907 -0.424540949  0.09723494 -1.015714095
#>  [60,] -1.39553199  0.069832439 -0.573475123  1.57110721 -0.198333888
#>  [61,]  0.68909899  0.898194495 -1.467699263  1.27930370 -0.183506201
#>  [62,]  0.55477636 -0.702992245  0.938556796  1.21211411 -0.501261979
#>  [63,]  0.32201899 -0.380250185 -1.513099576  0.95548762 -0.662709970
#>  [64,] -0.07013433  0.515997626 -0.196807834  1.44124858 -0.131861382
#>  [65,]  0.96885387 -0.773380361  1.717601700  1.09394026  1.020293175
#>  [66,]  0.88770974 -0.129729459  0.383455694 -0.65995066  1.040967404
#>  [67,]  0.17744732  0.359107987 -0.191037246  1.01359535 -0.008015303
#>  [68,]  0.10402927  0.778993245  1.540750266 -0.10967469 -0.424493016
#>  [69,]  2.08821108 -0.845179447 -0.032112567  0.29140756  2.157368340
#>  [70,]  0.46392520  2.390607889 -0.533046657 -1.34599363 -0.631509964
#>  [71,]  1.60737648  0.012340827  0.968631251 -0.79244226  0.779522254
#>  [72,] -0.24570954 -0.391322368  2.064741516  1.21674734  0.863258183
#>  [73,]  0.14815967  1.195069725 -0.436257096  1.91265155  2.332693787
#>  [74,]  0.24058334 -0.996637410 -0.432085483  0.50909714 -1.414356064
#>  [75,] -0.31999530  0.875249685  0.152000151  0.72132214  0.261091066
#>  [76,]  0.54585339  1.604698444  2.758852172  0.92635834  0.156980280
#>  [77,] -1.15626179 -0.920968096 -1.265940989  0.49775798 -0.242337767
#>  [78,]  0.24878370  1.495158252  1.642979126  0.11499351  2.908203800
#>  [79,] -0.68494245 -0.904793090 -1.002170178  0.49984891 -1.124140970
#>  [80,] -1.21985584  2.343513221  0.276717617  1.41704262 -0.949059403
#>  [81,]  0.21256668 -0.323965745 -0.452665697  1.72406519  1.155084421
#>  [82,] -0.15453176  0.393294931  0.236375186 -0.48231597 -0.357368180
#>  [83,]  0.86666033  0.833744051 -0.398803401  1.49589610  0.679033297
#>  [84,] -0.85080612 -0.007967635 -0.180273136  0.54055689  0.206341595
#>  [85,] -2.12752496 -0.740241570 -1.907797019  0.74768704  1.721947163
#>  [86,] -0.93261573  0.653814556 -1.051785800 -0.03738077  1.052366887
#>  [87,] -0.61048918 -1.130636176  0.972835390  1.44260248  0.782231441
#>  [88,]  0.18422505  0.410170430  0.740126076  1.41896562 -0.618604530
#>  [89,]  0.35702834 -0.538299030 -0.774245586  0.53359880 -0.414052243
#>  [90,]  1.70740317  0.762179800 -0.828024121  0.28266105  0.100106814
#>  [91,] -0.04738869  0.591006551 -0.059185939 -0.97366913 -1.380331648
#>  [92,]  1.77308252 -0.002046418  0.339560538 -1.49957728 -0.318530944
#>  [93,] -0.68673587 -0.922551346 -1.152971531  2.59729936  0.351108889
#>  [94,]  0.30431615  2.317815419  0.685118530 -0.48578698 -1.224281676
#>  [95,] -0.11567630  0.907554100  0.398078258  0.45171132  0.021818563
#>  [96,] -1.05733634  0.714524858 -1.210010050 -0.11593723 -0.855056890
#>  [97,]  1.70402015  1.125623216 -0.096523160  0.37470793  0.325343192
#>  [98,]  0.87339974 -0.732240453  0.188969509  0.62195253  2.191487962
#>  [99,] -0.38287282 -0.567284729  0.748502506  1.24650683  0.779208624
#> [100,] -1.51880478  0.121831358  0.972661802 -1.06352880  1.156957365
#>               [,6]         [,7]
#>   [1,] -0.99054628  1.038526790
#>   [2,] -0.61455836 -0.606917617
#>   [3,]  0.50766425  0.377150661
#>   [4,] -0.16114917 -0.338897752
#>   [5,] -0.09590246 -0.621870992
#>   [6,] -0.11908726 -1.616077833
#>   [7,]  1.76756360  0.085521186
#>   [8,]  0.27613640 -0.734374052
#>   [9,] -0.67378456  0.600035306
#>  [10,] -0.82261120  0.076270983
#>  [11,]  0.16103091  1.762362523
#>  [12,] -2.10817054  0.895485499
#>  [13,]  1.05741027  0.008974822
#>  [14,]  0.17429824 -0.764858191
#>  [15,]  1.05526845 -0.218443204
#>  [16,] -0.39175416 -0.405011537
#>  [17,]  0.59229038 -0.658874562
#>  [18,] -0.14579501 -0.060071277
#>  [19,] -0.64693160 -0.497152776
#>  [20,] -0.80153455 -0.248430434
#>  [21,] -1.28269935 -0.288091855
#>  [22,] -1.04474889 -0.455692304
#>  [23,]  1.10925850  1.215077929
#>  [24,] -0.80892031  0.283965822
#>  [25,]  2.05307183  0.303743972
#>  [26,]  1.36523355 -0.609686928
#>  [27,]  0.85612228  0.891229043
#>  [28,]  1.94530808 -0.119868069
#>  [29,]  0.07472336  3.383608882
#>  [30,] -0.41622622  1.382367273
#>  [31,]  0.41350442 -1.309253921
#>  [32,] -0.64812571  0.059281332
#>  [33,] -0.21397480 -0.539566300
#>  [34,] -1.02548799  0.168674202
#>  [35,] -1.28020380 -0.031580882
#>  [36,]  1.03387415 -1.428449378
#>  [37,]  1.23895857 -0.955603712
#>  [38,]  0.59043526  0.782916171
#>  [39,]  0.46782133  1.467968036
#>  [40,] -0.69414697  0.113080728
#>  [41,] -0.92154132 -1.040442683
#>  [42,]  1.38802330  0.730732977
#>  [43,] -0.23150889 -1.050296959
#>  [44,]  0.64325530 -0.579885743
#>  [45,]  1.04530070 -0.739403352
#>  [46,]  0.63313597  1.423598366
#>  [47,]  0.28448554  0.124405262
#>  [48,]  0.04501506  1.884731812
#>  [49,]  0.21459857  0.302412667
#>  [50,] -0.51314429 -0.370680002
#>  [51,]  0.68603953  0.456529618
#>  [52,] -0.93453299  0.156186173
#>  [53,] -0.13053183  1.285475579
#>  [54,] -0.73329679  0.355039590
#>  [55,]  1.09353922  0.495339548
#>  [56,] -0.85006782  0.302458669
#>  [57,] -1.53667567  0.398349297
#>  [58,]  0.27329316  0.399752280
#>  [59,]  1.12343315  0.215958851
#>  [60,] -2.19769056  0.157266591
#>  [61,]  0.90309251 -2.445870176
#>  [62,]  0.32095816  1.370113317
#>  [63,] -1.19086062  0.486126532
#>  [64,]  0.32008716 -0.382765527
#>  [65,]  1.26401542 -0.971008376
#>  [66,]  1.04260148  0.733609447
#>  [67,] -0.26544885 -1.915765331
#>  [68,]  1.39192422 -0.374721476
#>  [69,] -0.57945078 -0.378352026
#>  [70,] -1.52354812  0.390368580
#>  [71,]  0.81399873 -0.817822422
#>  [72,] -0.43275402  0.432092561
#>  [73,]  1.18993521 -0.012737919
#>  [74,]  1.16855516 -0.076867817
#>  [75,] -0.17689465  0.518616129
#>  [76,]  0.94780696  0.244025178
#>  [77,]  0.83603123  2.091290300
#>  [78,]  0.84175265  0.297508864
#>  [79,] -1.44911022  0.377138225
#>  [80,]  1.19006988 -0.208379836
#>  [81,]  0.68921967  0.398579692
#>  [82,]  1.23210852 -0.131069088
#>  [83,] -0.37236339 -0.447856473
#>  [84,]  0.25212546 -1.036599057
#>  [85,] -0.04625765  2.246925379
#>  [86,]  0.60426859  0.042591518
#>  [87,] -0.84678250  0.281207105
#>  [88,]  1.39902921  1.346111893
#>  [89,]  0.82477873  0.546276942
#>  [90,] -0.34276125 -0.488954784
#>  [91,]  0.15489358 -0.086998634
#>  [92,]  0.53597592  1.224240192
#>  [93,]  0.85256054 -0.320834287
#>  [94,] -0.09377581  0.071712193
#>  [95,]  1.28829819  0.206471942
#>  [96,]  3.46282834  0.258355534
#>  [97,]  0.16778501  1.654417552
#>  [98,]  0.37712647  0.297370231
#>  [99,]  1.42764520  0.648500447
#> [100,] -0.40120723 -0.058631752

pmwg:::particle_draws(0, rep(0.2, 7), diag(rep(7)))

#> NULL