Update the subject specific scaling parameters (epsilon)

Update the subject specific scaling parameter according to procedures outlined in P. H. Garthwaite, Y. Fan & S. A. Sisson (2016) Adaptive optimal scaling of Metropolis–Hastings algorithms using the Robbins–Monro process, Communications in Statistics - Theory and Methods, 45:17, 5098-5111, DOI: 10.1080/03610926.2014.936562

update_epsilon(epsilon, acc, p, i, d, alpha)

Arguments

epsilon

The scaling parameter for all subjects

acc

A boolean vector, TRUE if current sample != last sample

p

The target sample acceptance rate (0-1)

i

The current iteration for sampling

d

The number of parameters for the model

alpha

A hyperparameter for the epsilon tuning

Value

A vector with the new subject specific epsilon values