Test the arguments to the run_stage function for correctness

Takes the arguments to run_stage and checks them for completeness and correctness. Returns a list of cleaned/checked arguments to the caller.

check_run_stage_args(
  pmwgs,
  stage,
  iter = 1000,
  particles = 1000,
  display_progress = TRUE,
  n_cores = 1,
  ...
)

Arguments

pmwgs
A Particle Metropolis within Gibbs sampler which has been set up and initialised

stage
The sampling stage to run. Must be one of 'burn', 'adapt' or 'sample'.

iter
The number of iterations to run for the sampler. For 'burn' and 'sample' all iterations will run. However for 'adapt' if all subjects have enough unique samples to create the conditional distribution then the stage will finish early.

particles
The default here is 1000 particles to be generated for each iteration, however during the sample phase this should be reduced.

display_progress
Display a progress bar during sampling.

n_cores
Set to more than 1 to use mclapply. Setting n_cores greater than 1 is only permitted on Linux and Mac OS X machines.

...
Further arguments used to fine tune the sampling process. Accepted additional arguments include:

epsilon which should be a value between 0 and 1. epsilon controls the extent to which the covariance matrix is scaled when generating particles from the previous random effect.

mix controls the mixture of different sources for particles. The function numbers_from_proportion is passed this value and includes extra details on what is accepted.

n_unique is a number representing the number of unique samples to check for on each iteration of the sampler. Only used during the 'adapt' stage.